backend/requirements.txt
Django==5.1.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
openai==1.51.0
requests==2.32.3